+++ Changelog.md
@@ -2,6 +2,7 @@
 
 - New: Randomize signature button for the Keycard Reader
 - Change: Several technical blocks' sounds have been adjusted to better match how they look
+- Change: Improved visuals when holding a taser
 - Fix: Entities around a player sometimes get invisible when the player dismounts a camera
 - Fix: Shields do not take damage when blocking a taser that deals high enough damage
 - Fix: The death message of a player killed by a taser does not contain the player firing the taser